RAGfromWEB/requirements.txt
beautifulsoup4
feedparser
mattsollamatools
newspaper3k
nltk
numpy
requests
scikit_learn
sentence_transformers